% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-factorisation.R, R/plot-mf.R
\name{matrix_factorisation}
\alias{matrix_factorisation}
\alias{plot.mf}
\title{Matrix factorisation}
\usage{
matrix_factorisation(Y, k_dim, test = NULL, epochs = 500, lr = 0.001,
  lambda = 0.02, tol = 1e-06, pbar = FALSE)

\method{plot}{mf}(x, ...)
}
\arguments{
\item{Y}{Matrix with missing values. Missing values should be \code{NA}}

\item{k_dim}{Number of latent variables}

\item{test}{The select test. See \code{select_test}}

\item{epochs}{Maximum iterations}

\item{lr}{Learning rate}

\item{lambda}{Regularisation parameter}

\item{tol}{Iteration tolerance}

\item{pbar}{Progress bar toggle}

\item{x}{Matix factorisation object \code{mf}}

\item{...}{Dots}
}
\description{
Applies matrix factorisation by gradient descent (Funk SVD) to a given matrix with missing values

Plots the convergence of the matrix factorisation
}
\details{
Implements matrix factorisation by gradient descent. Matrix \code{Y} should have \code{NA}'s as missing values. The matrix will be factorised
into two matrices U and V, the user and feature matrix. The U matrix essentially contains weight each user gives to a certain feature. The
function will output the accuracy of the selected test sample by using \code{test}. See \code{select_test} for more information.

Plots the convergence profile.
}
\examples{
\dontrun{
m <- matrix(sample(c(NA, 1:5), 60, replace = TRUE, prob = c(0.2, rep(0.8/5, 5))), nrow = 10)
id <- select_test(m, 0.2)
mf <- matrix_factorisation(m, 2, test = id$test)
mf$pred
}
\dontrun{
m <- matrix(sample(c(NA, 1:5), 60, replace = TRUE, prob = c(0.2, rep(0.8/5, 5))), nrow = 10)
id <- select_test(m, 0.2)
mf <- matrix_factorisation(m, 2, test = id$test)
plot(mf)
}
}
