% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-test.R
\name{select_test}
\alias{select_test}
\title{Matrix factorisation test set}
\usage{
select_test(Y, r, min_row = 1)
}
\arguments{
\item{Y}{Matrix with missing values. Missing values should be \code{NA}}

\item{r}{Ratio i.e. proption of observations held out for testing}

\item{min_row}{The minimum observations to leave in each row. This protects against leaving an entire row missing.}
}
\description{
Select a test set to pass to \code{matrix_factorisation}
}
\details{
Selects a test to be used with \code{matrix_factorisation}. It returns the train set - the original matrix with NA's for cells
held out for the test and the test id vector. This can be useful to monitor the convergence and prediction accuracy.
}
\examples{
\dontrun{
m <- matrix(sample(c(NA, 1:5), 60, replace = TRUE, prob = c(0.2, rep(0.8/5, 5))), nrow = 10)
id <- select_test(m, 0.2)
mf <- matrix_factorisation(m, 2, test = id$test)
mf$pred
}
}
