% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/als.R
\name{als}
\alias{als}
\title{Alternating Least Squares}
\usage{
als(Y, k_dim, test = NULL, epochs = 200, lambda = 0.01,
  tol = 1e-06, pbar = FALSE)
}
\arguments{
\item{Y}{Matrix with missing values. Missing values should be \code{NA}}

\item{k_dim}{Number of latent variables}

\item{test}{The select test. See \code{select_test}}

\item{epochs}{Maximum iterations}

\item{lambda}{Regularisation parameter}

\item{tol}{Iteration tolerance}

\item{pbar}{Progress bar toggle}
}
\description{
Implements alternating least squares matrix factorisation on a given matrix
}
\details{
Implements Alternating least squares. Matrix \code{Y} should have \code{NA}'s as missing values. The matrix will be factorised
into two matrices U and V, the user and feature matrix. The U matrix essentially contains weight each user gives to a certain feature. The
function will output the accuracy of the selected test sample by using \code{test}. See \code{select_test} for more information.
}
\examples{
\dontrun{
m <- matrix(sample(1:5, 60, replace = TRUE), nrow = 10)
mf <- als(m, 2)
mf$pred
}
}
